server/requirements.txt
aiodns==3.0.0
aiohttp==3.8.5
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
ccxt==4.0.81
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
exceptiongroup==1.2.2
fastapi==0.105.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
loguru==0.7.0
multidict==6.0.4
mysql-connector-python==8.1.0
protobuf==4.21.12
psutil==5.9.6
pyaes==1.6.1
pyasn1==0.5.0
pycares==4.3.0
pycparser==2.21
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.27.0
Telethon==1.31.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.4
uvicorn==0.24.0.post1
yarl==1.9.2
